# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Music Generator API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GenerateRequest {
  /** Text description of the music to generate */
  prompt: string;
}

export type TaskStatusStatus =
  (typeof TaskStatusStatus)[keyof typeof TaskStatusStatus];

export const TaskStatusStatus = {
  pending: "pending",
  waking: "waking",
  generating: "generating",
  complete: "complete",
  error: "error",
} as const;

export interface TaskStatus {
  taskId: string;
  status: TaskStatusStatus;
  prompt: string;
  resultUrl?: string | null;
  error?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

/**
 * Opaque session token — Bearer <sid>.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};