# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Music Generator API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /generate:
    post:
      operationId: generateMusic
      tags: [music]
      summary: Start music generation
      description: Accepts a prompt and kicks off async background music generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRequest"
      responses:
        "202":
          description: Task queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStatus"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /status/{taskId}:
    get:
      operationId: getTaskStatus
      tags: [music]
      summary: Get task status
      description: Returns the current status and result of a music generation task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStatus"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    GenerateRequest:
      type: object
      properties:
        prompt:
          type: string
          description: Text description of the music to generate
      required:
        - prompt
    TaskStatus:
      type: object
      properties:
        taskId:
          type: string
        status:
          type: string
          enum: [pending, waking, generating, complete, error]
        prompt:
          type: string
        resultUrl:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - taskId
        - status
        - prompt
        - createdAt
        - updatedAt
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...