# lib/api-zod/src/generated/types/generateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Music Generator API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateRequest {
  /** Text description of the music to generate */
  prompt: string;
}