# lib/api-zod/src/generated/types/taskStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Music Generator API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskStatusStatus } from "./taskStatusStatus";

export interface TaskStatus {
  taskId: string;
  status: TaskStatusStatus;
  prompt: string;
  resultUrl?: string | null;
  error?: string | null;
  createdAt: Date;
  updatedAt: Date;
}